+++ README.md
@@ -26,7 +26,7 @@
 
 ### 차
 - [x] 가로 세로로 끝까지 갈 수 있다.
-- [x] 이동 경로에 기물이 있으면 움직일 수 없다.
+- [x] 이동 경로에 적군 기물이 있으면 움직일 수 없다.
 
 ### 사
 - [ ] 궁성 안에서만 움직일 수 있다.
