+++ src/main/java/janggi/domain/piece/behavior/Cannon.java
@@ -1,9 +1,10 @@
-package janggi.domain.piece;
+package janggi.domain.piece.behavior;
 
 import janggi.domain.Board;
-import janggi.domain.Position;
+import janggi.domain.move.Position;
 import janggi.domain.Side;
-import janggi.domain.Vector;
+import janggi.domain.move.Vector;
+import janggi.domain.piece.PieceBehavior;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
