+++ src/main/java/janggi/domain/piece/behavior/Guard.java
@@ -1,17 +1,21 @@
-package janggi.domain.piece;
+package janggi.domain.piece.behavior;
 
 import janggi.domain.Board;
-import janggi.domain.Position;
+import janggi.domain.move.Position;
 import janggi.domain.Side;
-import janggi.domain.Vector;
+import janggi.domain.move.Vector;
+import janggi.domain.piece.PieceBehavior;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class Soldier implements PieceBehavior {
+public class Guard implements PieceBehavior {
 
-    private static final List<Vector> VECTORS = List.of(new Vector(1, 0), new Vector(0, -1), new Vector(0, 1));
+    private static final List<Vector> VECTORS = List.of(
+            new Vector(1, 0), new Vector(0, -1), new Vector(0, 1), new Vector(-1, 0),
+            new Vector(1, 1), new Vector(-1, -1), new Vector(1, -1), new Vector(-1, 1)
+    );
 
     @Override
     public Set<Position> generateAvailableMovePositions(Board board, Side side, Position position) {
@@ -24,6 +28,6 @@ public class Soldier implements PieceBehavior {
 
     @Override
     public String toName() {
-        return "병";
+        return "사";
     }
 }
