+++ src/main/java/janggi/manager/JanggiGame.java
@@ -1,7 +1,7 @@
 package janggi.manager;
 
 import janggi.domain.Board;
-import janggi.domain.Position;
+import janggi.domain.move.Position;
 import janggi.domain.Side;
 import janggi.dto.PositionDto;
 import janggi.util.RecoveryUtil;
