+++ src/test/java/janggi/domain/VectorTest.java
@@ -2,6 +2,8 @@ package janggi.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.domain.move.Vector;
+import janggi.domain.move.Vectors;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
