+++ src/test/java/janggi/domain/piece/behavior/ChariotTest.java
@@ -1,11 +1,11 @@
-package janggi.domain.piece;
+package janggi.domain.piece.behavior;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.Board;
-import janggi.domain.Piece;
-import janggi.domain.Position;
+import janggi.domain.move.Position;
 import janggi.domain.Side;
+import janggi.domain.piece.Piece;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
